requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-filter>=23.5
drf-spectacular>=0.27
python-dotenv>=1.0
gunicorn>=21.2
Pillow
